(WIP) Refactor into multiple FSMs per connection
Change event structure in the event queue to represent an awaited I/O
event (e.g. read, send, recv, ...) corredated by some user data

Add CompletedEvent struct to represent a completed I/O event with a
return value and the user data passed when the event was added to the
queue

Make the EventQueue generic on T and explicitly take care of the
lifecycle of events

Change the addAsyncEvent method to work with the new event type

Change EventQueue.next to return a completed event

Create a fsm module to represent finite state machines for various
interactions (e.g. server gets a new connection, a signal is sent to the
program, some data arrived on an open connection, etc.). WIP

Add a Server FSM to represent the server process waiting for connections

Add a Connection FSM to represent an open connection and flows between
the server and the peer

Add a FSM union to carry a FSM and a pointer to some globally shared
data among multiple FSMs (or other parts of the program even)

Add explanatory comments for various sections of main to be refactored
at a later time

Change initialization code to work with the new data structures and
model

Reimplement the event loop with a single generic flow (client opens a
connection, sends a command, gets back a response) to verify
functionality and ergonmics. WIP

Make a util module for useful functions for every other module