(WIP) FSM Refactoring
Add a doubly linked list of commands to execute to the Connection FSM

Add a semaphore eventfd to the Connection FSM to be notified when new
commands are available on the queue

Add a notify event to event_queue.zig to write to a sempahore event fd
to post a notification

Add the hashset of slaves to the global data shared across FSMs

Refactor main to reduce repetition (e.g. when sending a response,
transitioning to waiting commands, etc.)

Allocate all FSM on the heap for easier cleanup at exit

Separate parsing of bytes on the wire into commands and execution of
those commands in two different states, using the notification eventfd
to switch from one to the other

Add support for passing an allocator to destruct user data in
EventQueue.destroy

Add an encode method to Command to allow writing one on the wire (e.g.
for propagation)

Add a destroy method to Command to allow cleaning up of allocated slices
when done

Add a format method to StreamRangeBound for easier formatting (e.g. in
command encoding)

Add a format method to StreamEntryID for easier formatting (e.g. in
responses and command encoding)

Clean up fsm.zig a bit

Squash a few bugs here and there