feat: Add initial support for SUBSCRIBE command
This commit introduces the initial implementation of the SUBSCRIBE
command.

A new subscribed_to state has been added to the connection's finite
state machine. When a client issues a SUBSCRIBE command, the connection
transitions to this state.

In the subscribed_to state, the server will reject most commands other
than SUBSCRIBE.